+++ src/commands/manuals/build.rs
@@ -32,7 +32,7 @@ DESCRIPTION
     has defined arguments, the additional arguments passed from the command-line
     will follow the previously defined arguments.
     
-    The subprocess will spawn from the current working ip's root directory.
+    The subprocess will spawn from the current working ip's build directory.
 
 OPTIONS
     --plugin <name>
@@ -43,6 +43,9 @@ OPTIONS
 
     --list
         View available plugins
+
+    --force
+        Execute the command without checking for a blueprint
 
     --build-dir <dir>
         The relative directory to locate the blueprint file
@@ -57,4 +60,5 @@ EXAMPLES
     orbit build --plugin xsim -- --elab
     orbit build --command python -- synth.py
     orbit build --verbose
+    orbit build --plugin xsim --force -- --help
 "#;